third_party/blink/web_tests/external/wpt/html/webappapis/dynamic-markup-insertion/html-unsafe-methods/Document-parseHTMLUnsafe-url.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] valid: created normally
  assert_equals: document.baseURI expected "about:blank" but got "http://web-platform.test:8001/html/webappapis/dynamic-markup-insertion/html-unsafe-methods/Document-parseHTMLUnsafe-url.tentative.html"
[FAIL] valid: created using another iframe's parseHTMLUnsafe from this frame
  assert_equals: document.baseURI expected "about:blank" but got "http://web-platform.test:8001/html/webappapis/dynamic-markup-insertion/html-unsafe-methods/resources/parseHTMLUnsafe-iframe.html"
[FAIL] valid: created using another iframe's parseHTMLUnsafe from that frame
  assert_equals: document.baseURI expected "about:blank" but got "http://web-platform.test:8001/html/webappapis/dynamic-markup-insertion/html-unsafe-methods/resources/parseHTMLUnsafe-iframe.html"
[FAIL] invalid XML: created normally
  assert_equals: document.baseURI expected "about:blank" but got "http://web-platform.test:8001/html/webappapis/dynamic-markup-insertion/html-unsafe-methods/Document-parseHTMLUnsafe-url.tentative.html"
[FAIL] invalid XML: created using another iframe's parseHTMLUnsafe from this frame
  assert_equals: document.baseURI expected "about:blank" but got "http://web-platform.test:8001/html/webappapis/dynamic-markup-insertion/html-unsafe-methods/resources/parseHTMLUnsafe-iframe.html"
[FAIL] invalid XML: created using another iframe's parseHTMLUnsafe from that frame
  assert_equals: document.baseURI expected "about:blank" but got "http://web-platform.test:8001/html/webappapis/dynamic-markup-insertion/html-unsafe-methods/resources/parseHTMLUnsafe-iframe.html"
Harness: the test ran to completion.
